# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Weather API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get jwt token",
                "operationId": "bae68d1c55329ecb6e41f48285840c8f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+380951234567"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "phone": "+380951234567",
                                    "password": "88888888"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vd2VhdGhlci1hcHAudGVzdC9hcGkvdjEvbG9naW4iLCJpYXQiOjE2OTg2MTM1OTMsImV4cCI6MTY5ODYxNzE5MywibmJmIjoxNjk4NjEzNTkzLCJqdGkiOiI4elpHYnR1TVlwd01tWlgwIiwic3ViIjoiMSIsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.Xh8THd6tQIgHlQ2r8yoTqPQ0LZRghEej0dJm_sv5XKw",
                                        "token_type": "bearer",
                                        "expires_in": 360000,
                                        "user": {
                                            "id": 1,
                                            "name": "Alexis Gleason",
                                            "email": "fisher.nellie@example.com",
                                            "phone": "+380954474265",
                                            "city": "Florida",
                                            "email_verified_at": "2023-10-29T08:33:28.000000Z",
                                            "created_at": "2023-10-29T08:33:29.000000Z",
                                            "updated_at": "2023-10-29T08:33:29.000000Z"
                                        }
                                    },
                                    "code": 200
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "get authorized user",
                "operationId": "Get user info",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user/weather": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "get weather by city of user",
                "operationId": "GetWeatherByCityUser",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    }
}